main/requirements.txt
chardet==3.*
googletrans==3.0.0
img2pdf==0.5.1
numpy==1.26.4
pdf2image==1.17.0
Pillow==10.3.0
PyPDF2==3.0.1
RealESRGAN==0.3.0
torch==2.3.0
ttkbootstrap==1.10.1
